Move the legacy classes to namespace App\Legacy
